Fix compile for stricter compilers...

+++ src/server/game/Handlers/MiscHandler.cpp
@@ -1767,7 +1767,7 @@ void WorldSession::HandleInspectRatedBGStatsOpcode(WorldPacket& recvData)
     Player* player = ObjectAccessor::FindPlayer(guid);
     if (!player)
     {
-        sLog->outDebug(LOG_FILTER_NETWORKIO, "CMSG_REQUEST_INSPECT_RATED_BG_STATS: No player found from GUID: " UI64FMTD, guid);
+        sLog->outDebug(LOG_FILTER_NETWORKIO, "CMSG_REQUEST_INSPECT_RATED_BG_STATS: No player found from GUID: " UI64FMTD, uint64(guid));
         return;
     }
